refactor: delete document_service and tag_service facades
These were pure re-export files with zero logic — just indirection
between UI code and database/models. Updated 3 UI files to import
directly from the source modules.
